resume_parser/backend/requirements.txt
Flask==2.1.1
flask_sqlalchemy==2.5.1
flask_restful==0.3.9
pdfminer.six==20201018
docx2txt==0.8
psycopg2-binary==2.9.3
